refactor(Panels CardAccordion): Move to new file
