+++ src/components/Card/Panels/PanelHeader.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+export const PanelHeader = styled.div`
+  display: flex;
+  font-weight: bold;
+  justify-content: space-between;
+  align-items: center;
+`;
